ADD: get method smells before and after
In the logs: no refactoring presents an improvement in code smell.
So there is obviously a bug, either in my code or in my assumptions of what the fields should be… Or in the data …. But it’s impossible that *NEVER* not even one refactoring improves code smells.

Possible culprits / potential mistakes: in particular those variables, I’m relying on them, but I’m not sure if they really work as I’m assuming they do:

                Console.WriteLine("refactoring.SourceClassName ==> " + refactoring.SourceClassName);
                Console.WriteLine("refactoring.TargetClassName ==> " + refactoring.TargetClassName);
                Console.WriteLine("refactoring.SourceOperatationName ==> " + refactoring.SourceOperatationName);
                Console.WriteLine("refactoring.TargetOperatationName ==> " + refactoring.TargetOperatationName);