FIX: previous commit also included a bug fix, and now I see there are 0 commits improved and 10 worsened by refactorings, according to Opus/Organic
The bug was that I didn’t log the results properly. They were found but not logged.
The fact that it can find some cases of set differences is comforting.
However the result is contrary to expectations: devs’ refactorings make the code worse, according to Marinesc’s Opus/Organic code smell detection tool.